Commit #7. Refacing page complete, changes to business logic

+++ src/main/java/com/benwillcabinets/benwillestimator/domain/ProjectItem.java
@@ -4,35 +4,37 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
+import java.math.BigDecimal;
 
 @Entity
 public class ProjectItem {
     @Id   //add a primary key
     @GeneratedValue
     private Integer id;
-    private double costProjectPrice;
-    private double sellProjectPrice;
+    private BigDecimal costProjectPrice;
+    private BigDecimal sellProjectPrice;
     @ManyToOne(targetEntity=Product.class)
     private Product product;
     private double qty;
+
 
     public Integer getId() {
         return id;
     }
 
-    public double getCostProjectPrice() {
+    public BigDecimal getCostProjectPrice() {
         return costProjectPrice;
     }
 
-    public void setCostProjectPrice(double costProjectPrice) {
+    public void setCostProjectPrice(BigDecimal costProjectPrice) {
         this.costProjectPrice = costProjectPrice;
     }
 
-    public double getSellProjectPrice() {
+    public BigDecimal getSellProjectPrice() {
         return sellProjectPrice;
     }
 
-    public void setSellProjectPrice(double sellProjectPrice) {
+    public void setSellProjectPrice(BigDecimal sellProjectPrice) {
         this.sellProjectPrice = sellProjectPrice;
     }
